Upstream

+++ code/game/objects/structures/watercloset.dm
@@ -266,6 +266,8 @@
 	if(isturf(loc))
 		var/turf/tile = loc
 		for(var/obj/effect/E in tile)
+			if (istype(E,/obj/effect/decal/cleanable/puddle_chem))
+				continue
 			if(istype(E,/obj/effect/rune) || istype(E,/obj/effect/decal/cleanable) || istype(E,/obj/effect/overlay))
 				qdel(E)
 
@@ -288,7 +290,7 @@
 	if(!ismist && is_washing)
 		return
 	is_washing = 1
-	var/turf/T = get_turf(src)
+	var/turf/T = loc
 	reagents.splash(T, reagents.total_volume)
 	T.clean(src)
 	spawn(100)
